Refactor code for norminette
