+++ Makefile
@@ -23,7 +23,6 @@ SOURCES = $(SRC_DIR)/main.c \
 					$(SRC_DIR)/parser/operators.c \
 					$(SRC_DIR)/parser/parser.c \
 					$(SRC_DIR)/parser/redir.c \
-					$(SRC_DIR)/parser/utils.c \
 					$(SRC_DIR)/debug.c #! Temporal
 OBJ = $(SOURCES:%.c=%.o)
