+++ source/parser/parser.c
@@ -6,35 +6,52 @@
 /*   By: migugar2 <migugar2@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/25 02:07:43 by migugar2          #+#    #+#             */
-/*   Updated: 2025/08/07 20:22:29 by migugar2         ###   ########.fr       */
+/*   Updated: 2025/08/08 16:22:21 by migugar2         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
-// TODO: Probably can print error when occurred, save info about error in parser struct and change return value
-
-int	manage_error(t_parser *parser)
+int	is_redirtok(t_tok *tok)
 {
-	// TODO: Add info about tokens in syntax error, manage freeing memory
-	// TODO: Probably can manage error when occurred
-	free_tokens(&parser->cur);
-	if (parser->err == PARSER_ERR_SYNTAX)
-	{
-		ft_putstr_fd("minishell: syntax error\n", STDERR_FILENO);
+	if (tok == NULL)
 		return (0);
-	}
-	if (parser->err == PARSER_ERR_MEMORY)
-	{
-		ft_putstr_fd("minishell: memory allocation error\n", STDERR_FILENO);
-		return (1);
-	}
-	if (parser->err == PARSER_ERR_UNEXPECTED_EOF)
-	{
-		ft_putstr_fd("minishell: syntax error near unexpected token `newline`\n", STDERR_FILENO);
-		return (1);
-	}
-	return (0);
+	return (tok->type == T_INFILE || tok->type == T_OUTFILE
+		|| tok->type == T_HEREDOC || tok->type == T_APPEND);
+}
+
+void	consume_tok(t_parser *parser)
+{
+	t_tok	*next;
+
+	if (parser->cur == NULL)
+		return ;
+	next = parser->cur->next;
+	free_tok(&parser->cur);
+	parser->cur = next;
+}
+
+char	*get_text_tok(t_tok *tok)
+{
+	if (tok->type == T_AND_IF)
+		return ("&&");
+	else if (tok->type == T_OR_IF)
+		return ("||");
+	else if (tok->type == T_PIPE)
+		return ("|");
+	else if (tok->type == T_LPAREN)
+		return ("(");
+	else if (tok->type == T_RPAREN)
+		return (")");
+	else if (tok->type == T_INFILE)
+		return ("<");
+	else if (tok->type == T_OUTFILE)
+		return (">");
+	else if (tok->type == T_HEREDOC)
+		return ("<<");
+	else if (tok->type == T_APPEND)
+		return (">>");
+	return ("");
 }
 
 int	parse_ast(t_tok *tok, t_ast **out)
@@ -50,17 +67,3 @@ int	parse_ast(t_tok *tok, t_ast **out)
 			free_ast_cmd_parse(out), 1);
 	return (0);
 }
-
-/*
-echo 123 | grep 123 && (echo 123 | grep 123)
-	&&
-|				\
-v					v
-  |					fork
-|   \					|
-v      v			v
-echo  grep			|
-							|		\
-							v			v
-						echo		grep
-*/
